+++ demos/basic-Vertex/index.js
@@ -1,0 +1,40 @@
+/**
+ * A script for demonstrating the basic usage of the Vertex class.
+ *
+ * @requires PlotBoilerplate, gup, dat.gui, 
+ * 
+ * @author   Ikaros Kappler
+ * @date     2020-05-18
+ * @version  1.0.0
+ **/
+
+
+(function(_context) {
+    "use strict";
+
+    // Fetch the GET params
+    let GUP = gup();
+
+    window.addEventListener(
+	'load',
+	function() {
+	    // All config params except the canvas are optional.
+	    var pb = new PlotBoilerplate(
+		PlotBoilerplate.utils.safeMergeByKeys(
+		    { canvas                : document.getElementById('my-canvas'),					    
+		      fullSize              : true
+		    }, GUP
+		)
+	    );
+
+	    // Create a new vertex
+	    var vertex = new Vertex( 50, 50 );
+
+	    // Now add it to your canvas
+	    pb.add( vertex );
+
+	    // Note: it's draggable now :)
+
+	} );
+    
+})(window);